fix(canary): move qr_scanner.h include after build_config.h
FEATURE_QR_PROVISION is defined in build_config.h, but the #include
guard for qr_scanner.h was placed before build_config.h was included.
The macro was always undefined at include time, so the scanner namespace
was never declared — but later code blocks (after build_config.h) tried
to use it. Move the include to after build_config.h.

Also fix CodeQL int overflow warning in quirc.c by casting to size_t
before multiplication.

+++ firmware/projects/canary-wap/arduino/canary_wap/quirc.c
@@ -78,8 +78,8 @@ int quirc_resize(struct quirc *q, int w, int h)
 
 	/* compute the "old" (i.e. currently allocated) and the "new"
 	   (i.e. requested) image dimensions */
-	size_t olddim = q->w * q->h;
-	size_t newdim = w * h;
+	size_t olddim = (size_t)q->w * (size_t)q->h;
+	size_t newdim = (size_t)w * (size_t)h;
 	size_t min = (olddim < newdim ? olddim : newdim);
 
 	/*
